refactor: unify init file location and names
