+++ src/commands/init/index.tsx
@@ -2,10 +2,10 @@ import type { CommandModule } from 'yargs';
 import * as React from 'react';
 import { render } from 'ink';
 import { ThemeProvider } from '@inkjs/ui';
-import { theme } from '../../components/theme.js';
+import { inkTheme } from '../../theme/ink-theme.js';
 import { checkIfConfigExists } from '../../config-file.js';
 import * as output from '../../output.js';
-import { InitUi } from './init-ui.js';
+import { InitUi } from './ui/InitUi.js';
 
 export const command: CommandModule<{}> = {
   command: ['init'],
@@ -17,7 +17,7 @@ export function run() {
   try {
     const hasConfig = checkIfConfigExists();
     render(
-      <ThemeProvider theme={theme}>
+      <ThemeProvider theme={inkTheme}>
         <InitUi hasConfig={hasConfig} />
       </ThemeProvider>,
     );
