+++ src/commands/prompt/ui/list/ProgramOutputListItem.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Text } from 'ink';
-import { colors } from '../../../../components/colors.js';
+import { colors } from '../../../../theme/colors.js';
 import type { ProgramOutputItem } from '../types.js';
 
 type ProgramOutputListItemProps = {
